+++ src/EPPlus.DataExtractor/DataExtractors/CollectionColumn/ICollectionColumnDataExtractor.cs
@@ -1,0 +1,10 @@
+﻿namespace EPPlus.DataExtractor.DataExtractors.CollectionColumn
+{
+    using OfficeOpenXml;
+
+    internal interface ICollectionColumnDataExtractor<in TRow>
+        where TRow : class, new()
+    {
+        void SetPropertyValue(TRow dataInstance, int row, ExcelRange cellRange);
+    }
+}
