+++ src/EPPlus.DataExtractor/DataExtractors/ColumnToCollectionDataExtractor.cs
@@ -1,5 +1,6 @@
-﻿namespace EPPlus.DataExtractor
+﻿namespace EPPlus.DataExtractor.DataExtractors
 {
+    using EPPlus.DataExtractor.DataExtractors.CollectionColumn;
     using OfficeOpenXml;
     using System;
     using System.Collections.Generic;
